# Trivia-Nexus/Trivia-Nexus/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quizara API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/start:
    post:
      operationId: startQuiz
      tags: [quiz]
      summary: Start a new quiz session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartQuizBody'
      responses:
        '201':
          description: Quiz session created with first question
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSession'
        '400':
          description: Invalid request
  /quiz/{sessionId}/answer:
    post:
      operationId: submitAnswer
      tags: [quiz]
      summary: Submit an answer for the current question
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitAnswerBody'
      responses:
        '200':
          description: Answer result with correct answer and explanation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerResult'
        '404':
          description: Session not found
  /quiz/{sessionId}:
    get:
      operationId: getQuizSession
      tags: [quiz]
      summary: Get a quiz session state
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Quiz session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizSession'
        '404':
          description: Session not found
  /quiz/{sessionId}/complete:
    post:
      operationId: completeQuiz
      tags: [quiz]
      summary: Complete a quiz session and get final results
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Final quiz results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResults'
...
components:
...
  schemas:
...
    StartQuizBody:
      type: object
      required: [categoryId]
      properties:
        categoryId:
          type: string
        questionCount:
          type: integer
          default: 10
        difficulty:
          type: [integer, 'null']
          description: 1=easy, 2=medium, 3=hard, null=mixed
    QuizQuestion:
      type: object
      required: [id, question, questionAr, options, optionsAr, questionNumber, totalQuestions]
      properties:
        id:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, image, fill_blank, ordering, matching, audio, hotspot]
          default: multiple_choice
        question:
          type: string
        questionAr:
          type: string
        imageUrl:
          type: [string, 'null']
        options:
          type: array
          items:
            type: string
        optionsAr:
          type: array
          items:
            type: string
        questionNumber:
          type: integer
        totalQuestions:
          type: integer
        difficulty:
          type: integer
    QuizSession:
      type: object
      required: [sessionId, categoryId, status, score, currentQuestion, questionNumber, totalQuestions]
      properties:
        sessionId:
          type: string
        categoryId:
          type: string
        categoryName:
          type: string
        status:
          type: string
          enum: [active, completed]
        score:
          type: integer
        currentQuestion:
          oneOf:
            - $ref: '#/components/schemas/QuizQuestion'
            - type: 'null'
        questionNumber:
          type: integer
        totalQuestions:
          type: integer
    SubmitAnswerBody:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: string
        selectedAnswer:
          type: integer
          minimum: 0
          description: Index of selected answer (MC, true_false, audio, image types)
        answerData:
          type: string
          description: JSON-encoded answer for complex types (ordering, matching, hotspot)
    AnswerResult:
      type: object
      required: [correct, correctAnswer, pointsEarned, coinsEarned, explanation, explanationAr, nextQuestion, sessionScore]
      properties:
        correct:
          type: boolean
        correctAnswer:
          type: integer
        correctAnswerData:
          type: [string, 'null']
          description: JSON-encoded correct answer for complex types (ordering, matching, hotspot)
        pointsEarned:
          type: integer
        coinsEarned:
          type: integer
        explanation:
          type: [string, 'null']
        explanationAr:
          type: [string, 'null']
        nextQuestion:
          oneOf:
            - $ref: '#/components/schemas/QuizQuestion'
            - type: 'null'
        sessionScore:
          type: integer
        isLastQuestion:
          type: boolean
    QuizResults:
      type: object
      required: [sessionId, score, totalQuestions, correctAnswers, timeTaken, categoryName]
      properties:
        sessionId:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        timeTaken:
          type: integer
          description: Time in seconds
        categoryName:
          type: string
        rank:
          type: [integer, 'null']
          description: Player's rank after this game
...